Move exercise helpers to dedicated module
