+++ tests/test_exercise_model.py
@@ -1,4 +1,5 @@
 import core
+from backend import exercises
 from backend.exercise import Exercise
 
 
@@ -14,7 +15,7 @@ def test_exercise_load_modify_save(sample_db):
         "scope": "set",
         "description": "",
     })
-    core.save_exercise(ex)
+    exercises.save_exercise(ex)
 
     loaded = Exercise("Push-up", db_path=sample_db, is_user_created=True)
     names = [m["name"] for m in loaded.metrics]
@@ -27,6 +28,6 @@ def test_had_metric(sample_db):
     assert ex.had_metric("Reps")
     ex.add_metric({"name": "Weight"})
     assert not ex.had_metric("Weight")
-    core.save_exercise(ex)
+    exercises.save_exercise(ex)
     loaded = Exercise("Push-up", db_path=sample_db, is_user_created=True)
     assert loaded.had_metric("Weight")
