+++ tests/test_workout_db.py
@@ -5,6 +5,7 @@ import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import core
+from backend import exercises
 
 
 def create_empty_db(path: Path) -> None:
@@ -95,17 +96,17 @@ def test_get_metric_type_schema(tmp_db: Path):
 
 
 def test_get_all_exercises_and_details(sample_db: Path):
-    all_ex = core.get_all_exercises(db_path=sample_db, include_user_created=True)
+    all_ex = exercises.get_all_exercises(db_path=sample_db, include_user_created=True)
     assert all_ex == [
         ("Bench Press", False),
         ("Push Up", False),
         ("Push Up", True),
     ]
-    details = core.get_exercise_details("Push Up", db_path=sample_db)
+    details = exercises.get_exercise_details("Push Up", db_path=sample_db)
     assert details["is_user_created"] is True
-    details_specific = core.get_exercise_details("Push Up", db_path=sample_db, is_user_created=False)
+    details_specific = exercises.get_exercise_details("Push Up", db_path=sample_db, is_user_created=False)
     assert details_specific["is_user_created"] is False
-    assert core.get_exercise_details("Nope", db_path=sample_db) is None
+    assert exercises.get_exercise_details("Nope", db_path=sample_db) is None
 
 
 def test_load_workout_presets(sample_db: Path):
