+++ ui/screens/exercise_library.py
@@ -19,6 +19,7 @@ from kivymd.uix.button import MDRaisedButton
 
 import os
 import core
+from backend import exercises
 from core import DEFAULT_DB_PATH
 
 
@@ -54,7 +55,7 @@ class ExerciseLibraryScreen(MDScreen):
             app and self.cache_version != getattr(app, "exercise_library_version", 0)
         ):
             db_path = DEFAULT_DB_PATH
-            self.all_exercises = core.get_all_exercises(
+            self.all_exercises = exercises.get_all_exercises(
                 db_path, include_user_created=True
             )
             if app:
@@ -104,7 +105,7 @@ class ExerciseLibraryScreen(MDScreen):
             app and self.cache_version != getattr(app, "exercise_library_version", 0)
         ):
             db_path = DEFAULT_DB_PATH
-            self.all_exercises = core.get_all_exercises(
+            self.all_exercises = exercises.get_all_exercises(
                 db_path, include_user_created=True
             )
             if app:
@@ -255,7 +256,7 @@ class ExerciseLibraryScreen(MDScreen):
         def do_delete(*args):
             db_path = DEFAULT_DB_PATH
             try:
-                core.delete_exercise(
+                exercises.delete_exercise(
                     exercise_name, db_path=db_path, is_user_created=True
                 )
                 app = MDApp.get_running_app()
