+++ ui/screens/rest_screen.py
@@ -54,7 +54,8 @@ except Exception:  # pragma: no cover - simple stubs
 
 import time
 import math
-from core import DEFAULT_REST_DURATION, get_exercise_details
+from core import DEFAULT_REST_DURATION
+from backend.exercises import get_exercise_details
 
 
 class RestScreen(MDScreen):
